+++ src/const.tsx
@@ -1,3 +1,4 @@
+import { OfferType } from './types.ts';
 
 export const cityMain = [
   {
@@ -26,42 +27,8 @@ export const cityMain = [
   },
 ];
 
-type Location = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-}
 
-type City = {
-  name: string;
-  location: Location;
-}
-
-type Host = {
-  name: string;
-  avatarUrl: string;
-  isPro: boolean;
-}
-
-export type Offer = {
-  id: string;
-  title: string;
-  type: string;
-  price: number;
-  city: City;
-  location: Location;
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-  description: string;
-  bedrooms: number;
-  goods: string[];
-  host: Host;
-  images: string[];
-  maxAdults: number;
-}
-
-const offers: Array<Offer> = [
+const offers: Array<OfferType> = [
   {
     id: '27cd0cb9-1380-4b6c-aabe-580dd05484b3',
     title: 'Wood and stone place',
